Separate HPAV-generic and Qualcomm vendor-specific constants and types
