+++ qualcomm.go
@@ -1,31 +1,12 @@
 package main
 
-import "net"
-
 const (
-	// HPAV EtherType
-	etherType = 0x88e1
-
-	hpVersion = 0
-
+	// Qualcomm vendor-specific Management Message types.
 	nwInfoReq = 0xa038
 	nwInfoCnf = 0xa039
 )
 
 var (
-	hpVendor = [...]byte{0x00, 0xb0, 0x52}
+	// Qualcomm Atheros OUI used for vendor-specific HPAV extensions.
+	ouiQualcomm = oui{0x00, 0xb0, 0x52}
 )
-
-// macAddr is a convenience type that is strictly limited to the Ethernet MAC address length,
-// unlike the standard library net.HardwareAddr type.
-type macAddr [6]byte
-
-func (m macAddr) String() string {
-	return net.HardwareAddr(m[:]).String()
-}
-
-type networkID [7]byte
-
-func (n networkID) String() string {
-	return net.HardwareAddr(n[:]).String()
-}
