+++ qualcomm_test.go
@@ -11,9 +11,9 @@ func TestHomeplugFrame(t *testing.T) {
 	buf := []byte{0x00, 0x39, 0xa0, 0x00, 0xb0, 0x52}
 
 	want := HomeplugFrame{
-		Version: 0,
+		Version: hpavVersion1_0,
 		MMEType: nwInfoCnf,
-		Vendor:  [3]byte{0x00, 0xb0, 0x52},
+		Vendor:  ouiQualcomm,
 		Payload: []uint8{},
 	}
